+++ test/adapters/soda_test.rb
@@ -14,4 +14,16 @@ class SodaTest < ActionDispatch::IntegrationTest
   def test_run
     assert_result [{"hello" => "world"}], "SELECT 'world' AS hello LIMIT 1"
   end
+
+  def test_single_quote
+    assert_result [{"hello" => "'"}], "SELECT {var} AS hello LIMIT 1", var: "'"
+  end
+
+  def test_double_quote
+    assert_result [{"hello" => '"'}], "SELECT {var} AS hello LIMIT 1", var: '"'
+  end
+
+  def test_backslash
+    assert_result [{"hello" => "\\"}], "SELECT {var} AS hello LIMIT 1", var: "\\"
+  end
 end